Updating the function and the sports team list
I changed the sports team lists to ensure all rosters had no missing heights values. This makes sure there are no errors when trying to convert the height values to inches. A dash in the height values would cause an error here, but because there no more missing height values this is no longer a problem.